refactor: 생성자 및 메서드 로직 개선
